correctly retrieve latest submission for programming exercise
